backend/requirements.txt
agno==2.2.13
openai>=2.8.0
python-dotenv>=1.0.0
pyyaml>=6.0
